feat: remove the typography type concept from existing slate code
