+++ packages/runtime/src/components/builtin/Text/components/Leaf/index.ts
@@ -1,2 +1,1 @@
 export * from './leaf'
-export * from './typography'
