+++ packages/runtime/src/controls/rich-text/index.ts
@@ -2,4 +2,3 @@ export * from './dto-types'
 export * from './rich-text'
 export * from './translation'
 export * from './types'
-export * from './compare'
